refactor : 토큰명 타입 -> 상수로 이동
