+++ apis/config/interceptor.server.ts
@@ -1,6 +1,7 @@
 import { ApiError } from './customError';
 import privateApi from './privateApi.server';
-import { AUTH_COOKIE_KEYS, type ErrorType } from './type';
+import { type ErrorType } from './type';
+import { AUTH_COOKIE_KEYS } from '@/constants';
 import { AUTH_ERROR_CODES } from '@/constants/errorCode';
 import { getAccessTokenServer } from '@/utils/auth/tokenValidator.server';
 import { AxiosError, type InternalAxiosRequestConfig } from 'axios';
