+++ apis/config/type.ts
@@ -27,9 +27,3 @@ export type ErrorType = {
   status: ErrorStatus;
   timestamp: Date;
 };
-
-export const AUTH_COOKIE_KEYS: Record<string, string> = {
-  accessToken: 'gloddy_at',
-  refreshToken: 'gloddy_rt',
-  userId: 'gloddy_uid',
-} as const;
